+++ skbeam/core/correlation.py
@@ -53,7 +53,7 @@ logger = logging.getLogger(__name__)
 
 def _one_time_process(buf, G, past_intensity_norm, future_intensity_norm,
                       label_array, num_bufs, num_pixels, img_per_level,
-                      level, buf_no):
+                      level, buf_no, norm, lev_len):
     """Reference implementation of the inner loop of multi-tau one time
     correlation
 
@@ -85,6 +85,10 @@ def _one_time_process(buf, G, past_intensity_norm, future_intensity_norm,
         the current multi-tau level
     buf_no : int
         the current buffer number
+    norm : dict
+        to track bad images
+    lev_len : array
+        length of each level
 
     Notes
     -----
@@ -99,21 +103,30 @@ def _one_time_process(buf, G, past_intensity_norm, future_intensity_norm,
     # in multi-tau correlation, the subsequent levels have half as many
     # buffers as the first
     i_min = num_bufs // 2 if level else 0
-
     for i in range(i_min, min(img_per_level[level], num_bufs)):
         # compute the index into the autocorrelation matrix
         t_index = level * num_bufs / 2 + i
-
         delay_no = (buf_no - i) % num_bufs
+
         # get the images for correlating
         past_img = buf[level, delay_no]
         future_img = buf[level, buf_no]
-        for w, arr in zip([past_img*future_img, past_img, future_img],
-                          [G, past_intensity_norm, future_intensity_norm]):
-            binned = np.bincount(label_array, weights=w)[1:]
-            # pdb.set_trace()
-            arr[t_index] += ((binned / num_pixels - arr[t_index]) /
-                             (img_per_level[level] - i))
+
+        # find the normalization that can work both for bad_images
+        #  and good_images
+        ind = int(t_index - lev_len[:level].sum())
+        normalize = img_per_level[level] - i - norm[level+1][ind]
+
+        # take out the past_ing and future_img created using bad images
+        # (bad images are converted to np.nan array)
+        if np.isnan(past_img).any() or np.isnan(future_img).any():
+            norm[level + 1][ind] += 1
+        else:
+            for w, arr in zip([past_img*future_img, past_img, future_img],
+                              [G, past_intensity_norm, future_intensity_norm]):
+                binned = np.bincount(label_array, weights=w)[1:]
+                arr[t_index] += ((binned / num_pixels -
+                                  arr[t_index]) / normalize)
     return None  # modifies arguments in place!
 
 
@@ -135,13 +148,14 @@ _internal_state = namedtuple(
      'pixel_list',
      'num_pixels',
      'lag_steps',
+     'norm',
+     'lev_len',
      ]
 )
 
 _two_time_internal_state = namedtuple(
     'two_time_correlation_state',
-    [
-     'buf',
+    ['buf',
      'img_per_level',
      'label_array',
      'track_level',
@@ -153,7 +167,9 @@ _two_time_internal_state = namedtuple(
      'count_level',
      'current_img_time',
      'time_ind',
-     ]
+     'norm',
+     'lev_len',
+    ]
 )
 
 
@@ -175,9 +191,9 @@ def _init_state_one_time(num_levels, num_bufs, labels):
         `lazy_one_time` requires so that it can be used to pick up
          processing after it was interrupted
     """
-    (label_array, pixel_list, num_rois,
-     num_pixels, lag_steps, buf, img_per_level, track_level,
-     cur) = _validate_and_transform_inputs(num_bufs, num_levels, labels)
+    (label_array, pixel_list, num_rois, num_pixels, lag_steps, buf,
+     img_per_level, track_level, cur, norm,
+     lev_len) = _validate_and_transform_inputs(num_bufs, num_levels, labels)
 
     # G holds the un normalized auto- correlation result. We
     # accumulate computations into G as the algorithm proceeds.
@@ -200,6 +216,8 @@ def _init_state_one_time(num_levels, num_bufs, labels):
         pixel_list,
         num_pixels,
         lag_steps,
+        norm,
+        lev_len,
     )
 
 
@@ -237,11 +255,12 @@ def lazy_one_time(image_iterable, num_levels, num_bufs, labels,
     ------
     namedtuple
         A `results` object is yielded after every image has been processed.
-         This `reults` object contains:
-        - the normalized correlation, `g2`
-        - the times at which the correlation was computed, `lag_steps`
-        - and all of the internal state, `final_state`, which is a
-          `correlation_state` namedtuple
+        This `reults` object contains, in this order:
+        - `g2`: the normalized correlation
+          shape is (len(lag_steps), num_rois)
+        - `lag_steps`: the times at which the correlation was computed
+        - `_internal_state`: all of the internal state. Can be passed back in
+          to `lazy_one_time` as the `internal_state` parameter
 
     Notes
     -----
@@ -290,7 +309,7 @@ def lazy_one_time(image_iterable, num_levels, num_bufs, labels,
         # and img_per_level in place!
         _one_time_process(s.buf, s.G, s.past_intensity, s.future_intensity,
                           s.label_array, num_bufs, s.num_pixels,
-                          s.img_per_level, level, buf_no)
+                          s.img_per_level, level, buf_no, s.norm, s.lev_len)
 
         # check whether the number of levels is one, otherwise
         # continue processing the next level
@@ -308,9 +327,7 @@ def lazy_one_time(image_iterable, num_levels, num_bufs, labels,
 
                 s.buf[level, s.cur[level] - 1] = ((
                         s.buf[level - 1, prev - 1] +
-                        s.buf[level - 1, s.cur[level - 1] - 1]
-                    ) / 2
-                )
+                        s.buf[level - 1, s.cur[level - 1] - 1]) / 2)
 
                 # make the track_level zero once that level is processed
                 s.track_level[level] = False
@@ -321,7 +338,8 @@ def lazy_one_time(image_iterable, num_levels, num_bufs, labels,
                 buf_no = s.cur[level] - 1
                 _one_time_process(s.buf, s.G, s.past_intensity,
                                   s.future_intensity, s.label_array, num_bufs,
-                                  s.num_pixels, s.img_per_level, level, buf_no)
+                                  s.num_pixels, s.img_per_level, level, buf_no,
+                                  s.norm, s.lev_len)
                 level += 1
 
                 # Checking whether there is next level for processing
@@ -346,7 +364,12 @@ def multi_tau_auto_corr(num_levels, num_bufs, labels, images):
     Original code(in Yorick) for multi tau auto correlation
     author: Mark Sutton
 
-    See docstring for lazy_one_time
+    For parameter description, please reference the docstring for
+    lazy_one_time. Note that there is an API difference between this function
+    and `lazy_one_time`. The `images` arugment is at the end of this function
+    signature here for backwards compatibility, but is the first argument in
+    the `lazy_one_time()` function. The semantics of the variables remain
+    unchanged.
     """
     gen = lazy_one_time(images, num_levels, num_bufs, labels)
     for result in gen:
@@ -363,14 +386,11 @@ def auto_corr_scat_factor(lags, beta, relaxation_rate, baseline=1):
     ----------
     lags : array
         delay time
-
     beta : float
         optical contrast (speckle contrast), a sample-independent
         beamline parameter
-
     relaxation_rate : float
         relaxation time associated with the samples dynamics.
-
     baseline : float, optional
         baseline of one time correlation
         equal to one for ergodic samples
@@ -389,10 +409,8 @@ def auto_corr_scat_factor(lags, beta, relaxation_rate, baseline=1):
         g_2(q, \tau) = \beta_1[g_1(q, \tau)]^{2} + g_\infty
 
     For a system undergoing  diffusive dynamics,
-
     :math ::
         g_1(q, \tau) = e^{-\gamma(q) \tau}
-
     :math ::
        g_2(q, \tau) = \beta_1 e^{-2\gamma(q) \tau} + g_\infty
 
@@ -415,7 +433,12 @@ def two_time_corr(labels, images, num_frames, num_bufs, num_levels=1):
     This function computes two-time correlation
     Original code : author: Yugang Zhang
 
-    See docstring for lazy_two_time
+    Returns
+    -------
+    results : namedtuple
+
+    For parameter definition, see the docstring for the `lazy_two_time()`
+    function in this module
     """
     gen = lazy_two_time(labels, images, num_frames, num_bufs, num_levels)
     for result in gen:
@@ -457,9 +480,14 @@ def lazy_two_time(labels, images, num_frames, num_bufs, num_levels=1,
 
     Yields
     ------
-    internal_state : tuple
-      all of the internal state, `final_state`, which is a
-      `correlation_state` namedtuple
+    namedtuple
+        A `results` object is yielded after every image has been processed.
+        This `reults` object contains, in this order:
+        - `g2`: the normalized correlation
+          shape is (num_rois, len(lag_steps), len(lag_steps))
+        - `lag_steps`: the times at which the correlation was computed
+        - `_internal_state`: all of the internal state. Can be passed back in
+          to `lazy_one_time` as the `internal_state` parameter
 
     Notes
     -----
@@ -505,7 +533,8 @@ def lazy_two_time(labels, images, num_frames, num_bufs, num_levels=1,
         # (undownsampled) frames. two_time and img_per_level in place!
         _two_time_process(s.buf, s.g2, s.label_array, num_bufs,
                           s.num_pixels, s.img_per_level, s.lag_steps,
-                          s.current_img_time, level=0, buf_no=s.cur[0] - 1)
+                          s.current_img_time,
+                          level=0, buf_no=s.cur[0] - 1)
 
         # time frame for each level
         s.time_ind[0].append(s.current_img_time)
@@ -546,8 +575,8 @@ def lazy_two_time(labels, images, num_frames, num_bufs, num_levels=1,
                 # on previous call above.
                 _two_time_process(s.buf, s.g2, s.label_array, num_bufs,
                                   s.num_pixels, s.img_per_level, s.lag_steps,
-                                  current_img_time, level=level,
-                                  buf_no=s.cur[level]-1)
+                                  current_img_time,
+                                  level=level, buf_no=s.cur[level]-1)
                 level += 1
 
                 # Checking whether there is next level for processing
@@ -570,15 +599,15 @@ def two_time_state_to_results(state):
         and the lag steps
     """
     for q in range(np.max(state.label_array)):
-        x0 = (state.g2)[:, :, q]
-        (state.g2)[:, :, q] = (np.tril(x0) + np.tril(x0).T -
+        x0 = (state.g2)[q, :, :]
+        (state.g2)[q, :, :] = (np.tril(x0) + np.tril(x0).T -
                                np.diag(np.diag(x0)))
     return results(state.g2, state.lag_steps, state)
 
 
 def _two_time_process(buf, g2, label_array, num_bufs, num_pixels,
-                      img_per_level, lag_steps, current_img_time, level,
-                      buf_no):
+                      img_per_level, lag_steps, current_img_time,
+                      level, buf_no):
     """
     Parameters
     ----------
@@ -586,6 +615,7 @@ def _two_time_process(buf, g2, label_array, num_bufs, num_pixels,
         image data array to use for two time correlation
     g2: array
         two time correlation matrix
+        shape (number of labels(ROI), number of frames, number of frames)
     label_array: array
         Elements not inside any ROI are zero; elements inside each
         ROI are 1, 2, 3, etc. corresponding to the order they are specified
@@ -594,7 +624,7 @@ def _two_time_process(buf, g2, label_array, num_bufs, num_pixels,
         number of buffers(channels)
     num_pixels : array
         number of pixels in certain ROI's
-        ROI's, dimensions are : [number of ROI's]
+        ROI's, dimensions are len(np.unique(label_array))
     img_per_level: array
         to track how many images processed in each level
     lag_steps : array
@@ -642,11 +672,11 @@ def _two_time_process(buf, g2, label_array, num_bufs, num_pixels,
         if not isinstance(current_img_time, int):
             nshift = 2**(level-1)
             for i in range(-nshift+1, nshift+1):
-                g2[int(tind1+i),
+                g2[:, int(tind1+i),
                    int(tind2+i)] = (tmp_binned/(pi_binned *
                                                 fi_binned))*num_pixels
         else:
-            g2[tind1, tind2] = tmp_binned/(pi_binned * fi_binned)*num_pixels
+            g2[:, tind1, tind2] = tmp_binned/(pi_binned * fi_binned)*num_pixels
 
 
 def _init_state_two_time(num_levels, num_bufs, labels, num_frames):
@@ -669,8 +699,8 @@ def _init_state_two_time(num_levels, num_bufs, labels, num_frames):
         after it was interrupted
     """
     (label_array, pixel_list, num_rois, num_pixels, lag_steps,
-     buf, img_per_level, track_level,
-     cur) = _validate_and_transform_inputs(num_bufs, num_levels, labels)
+     buf, img_per_level, track_level, cur, norm,
+     lev_len) = _validate_and_transform_inputs(num_bufs, num_levels, labels)
 
     # to count images in each level
     count_level = np.zeros(num_levels, dtype=np.int64)
@@ -682,7 +712,7 @@ def _init_state_two_time(num_levels, num_bufs, labels, num_frames):
     time_ind = {key: [] for key in range(num_levels)}
 
     # two time correlation results (array)
-    g2 = np.zeros((num_frames, num_frames, num_rois), dtype=np.float64)
+    g2 = np.zeros((num_rois, num_frames, num_frames), dtype=np.float64)
 
     return _two_time_internal_state(
         buf,
@@ -697,6 +727,8 @@ def _init_state_two_time(num_levels, num_bufs, labels, num_frames):
         count_level,
         current_img_time,
         time_ind,
+        norm,
+        lev_len,
     )
 
 
@@ -734,7 +766,11 @@ def _validate_and_transform_inputs(num_bufs, num_levels, labels):
     track_level : array
         to track processing each level
     cur : array
-
+        to increment the buffer
+    norm : dict
+        to track bad images
+    lev_len : array
+        length of each levels
     """
     if num_bufs % 2 != 0:
         raise ValueError("There must be an even number of `num_bufs`. You "
@@ -742,7 +778,8 @@ def _validate_and_transform_inputs(num_bufs, num_levels, labels):
     label_array, pixel_list = extract_label_indices(labels)
 
     # map the indices onto a sequential list of integers starting at 1
-    label_mapping = {label: n+1 for n, label in enumerate(np.unique(label_array))}
+    label_mapping = {label: n+1
+                     for n, label in enumerate(np.unique(label_array))}
     # remap the label array to go from 1 -> max(_labels)
     for label, n in label_mapping.items():
         label_array[label_array == label] = n
@@ -754,7 +791,12 @@ def _validate_and_transform_inputs(num_bufs, num_levels, labels):
     num_pixels = np.bincount(label_array)[1:]
 
     # Convert from num_levels, num_bufs to lag frames.
-    tot_channels, lag_steps = multi_tau_lags(num_levels, num_bufs)
+    tot_channels, lag_steps, dict_lag = multi_tau_lags(num_levels, num_bufs)
+
+    # these norm and lev_len will help to find the one time correlation
+    # normalization norm will updated when there is a bad image
+    norm = {key: [0] * len(dict_lag[key]) for key in (dict_lag.keys())}
+    lev_len = np.array([len(dict_lag[i]) for i in (dict_lag.keys())])
 
     # Ring buffer, a buffer with periodic boundary conditions.
     # Images must be keep for up to maximum delay in buf.
@@ -768,4 +810,30 @@ def _validate_and_transform_inputs(num_bufs, num_levels, labels):
     cur = np.ones(num_levels, dtype=np.int64)
 
     return (label_array, pixel_list, num_rois, num_pixels,
-            lag_steps, buf, img_per_level, track_level, cur)
+            lag_steps, buf, img_per_level, track_level, cur,
+            norm, lev_len)
+
+
+def one_time_from_two_time(two_time_corr):
+    """
+    This will provide the one-time correlation data from two-time
+    correlation data.
+
+    Parameters
+    ----------
+    two_time_corr : array
+        matrix of two time correlation
+        shape (number of labels(ROI's), number of frames, number of frames)
+
+    Returns
+    -------
+    one_time_corr : array
+        matrix of one time correlation
+        shape (number of labels(ROI's), number of frames)
+    """
+
+    one_time_corr = np.zeros((two_time_corr.shape[0], two_time_corr.shape[2]))
+    for g in two_time_corr:
+        for j in range(two_time_corr.shape[2]):
+            one_time_corr[:, j] = np.trace(g, offset=j)/two_time_corr.shape[2]
+    return one_time_corr