+++ skbeam/core/tests/test_correlation.py
@@ -44,7 +44,10 @@ from skbeam.core.correlation import (multi_tau_auto_corr,
                                      auto_corr_scat_factor,
                                      lazy_one_time,
                                      lazy_two_time, two_time_corr,
-                                     two_time_state_to_results)
+                                     two_time_state_to_results,
+                                     one_time_from_two_time)
+from skbeam.core.mask import bad_to_nan_gen
+
 
 logger = logging.getLogger(__name__)
 
@@ -169,7 +172,7 @@ def test_two_time_corr():
 
 def test_auto_corr_scat_factor():
     num_levels, num_bufs = 3, 4
-    tot_channels, lags = utils.multi_tau_lags(num_levels, num_bufs)
+    tot_channels, lags, dict_lags = utils.multi_tau_lags(num_levels, num_bufs)
     beta = 0.5
     relaxation_rate = 10.0
     baseline = 1.0
@@ -178,6 +181,45 @@ def test_auto_corr_scat_factor():
 
     assert_array_almost_equal(g2, np.array([1.5, 1.0, 1.0, 1.0, 1.0,
                                             1.0, 1.0, 1.0]), decimal=8)
+
+
+def test_bad_images():
+    setup()
+    g2, lag_steps = multi_tau_auto_corr(4, num_bufs,
+                                        rois, img_stack)
+    # introduce bad images
+    bad_img_list = [3, 21, 35, 48]
+    # convert each bad image to np.nan array
+    images = bad_to_nan_gen(img_stack, bad_img_list)
+
+    # then use new images (including bad images)
+    g2_n, lag_steps_n = multi_tau_auto_corr(4, num_bufs,
+                                            rois, images)
+
+    assert_array_almost_equal(g2[:, 0], g2_n[:, 0], decimal=3)
+    assert_array_almost_equal(g2[:, 1], g2_n[:, 1], decimal=3)
+
+
+def test_one_time_from_two_time():
+    num_lev = 1
+    num_buf = 10  # must be even
+    x_dim = 10
+    y_dim = 10
+    stack = 10
+    imgs = np.random.randint(1, 3, (stack, x_dim, y_dim))
+    roi = np.zeros_like(imgs[0])
+    # make sure that the ROIs can be any integers greater than 1.
+    # They do not have to start at 1 and be continuous
+    roi[0:x_dim//10, 0:y_dim//10] = 5
+    roi[x_dim//10:x_dim//5, y_dim//10:y_dim//5] = 3
+
+    g2, lag_steps, _state = two_time_corr(roi, imgs, stack,
+                                          num_buf, num_lev)
+
+    one_time = one_time_from_two_time(g2)
+    assert_array_almost_equal(one_time[0, :], np.array([1.0, 0.9, 0.8, 0.7,
+                                                        0.6, 0.5, 0.4, 0.3,
+                                                        0.2, 0.1]))
 
 
 if __name__ == '__main__':